Crate alloy_providers

alloy-providers

Ethereum JSON-RPC providers.

This crate contains the Provider trait, which exposes Ethereum JSON-RPC methods. Providers in alloy are similar to ethers.js Providers. They manage an RpcClient and allow other parts of the program to easily make RPC calls.

Unlike an ethers.js Provider, an alloy Provider is network-aware. It is parameterized with a Network from [alloy-networks]. This allows the Provider to expose a consistent interface to the rest of the program, while adjusting request and response types to match the underlying blockchain.

Providers can be composed via stacking. For example, a Provider that tracks the nonce for a given address can be stacked onto a Provider that signs transactions to create a Provider that can send signed transactions with correct nonces.

The ProviderBuilder struct can quickly create a stacked provider, similar to tower::ServiceBuilder.

Usage

TODO :)

Modules

provider

Alloy main Provider abstraction.

utils

Provider-related utilities.

Structs

NetworkRpcClient

A network-wrapped RPC client.

ProviderBuilder

A builder for constructing a Provider from various layers.

Stack

Traits

Provider

Provider is parameterized with a network and a transport. The default transport is type-erased, but you can do Provider<N, Http>.

ProviderLayer

A layering abstraction in the vein of tower::Layer