Struct alloy_providers::provider::Provider

pub struct Provider<T: Transport = BoxTransport> { /* private fields */ }

An abstract provider for interacting with the Ethereum JSON RPC API. Must be instantiated with a transport which implements the Transport trait.

Implementations

impl<T: Transport + Clone + Send + Sync> Provider<T>

pub fn new(transport: T) -> Self

pub fn new_with_client(client: RpcClient<T>) -> Self

pub fn with_sender(self, from: Address) -> Self

pub fn inner(&self) -> &RpcClient<T>