Trait alloy_providers::provider::TempProvider

pub trait TempProvider: Send + Sync {

    // Required methods
    fn get_transaction_count<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        tag: Option<BlockId>
    ) -> Pin<Box<dyn Future<Output = TransportResult<U256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_number<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = TransportResult<u64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_balance<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        tag: Option<BlockId>
    ) -> Pin<Box<dyn Future<Output = TransportResult<U256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_by_hash<'life0, 'async_trait>(
        &'life0 self,
        hash: BlockHash,
        full: bool
    ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Block>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_by_number<'life0, 'async_trait>(
        &'life0 self,
        number: BlockNumberOrTag,
        full: bool
    ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Block>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_client_version<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = TransportResult<String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_chain_id<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = TransportResult<U64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_net_version<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = TransportResult<U64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_storage_at<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        key: U256,
        tag: Option<BlockId>
    ) -> Pin<Box<dyn Future<Output = TransportResult<StorageValue>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_code_at<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        tag: BlockId
    ) -> Pin<Box<dyn Future<Output = TransportResult<Bytes>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_transaction_by_hash<'life0, 'async_trait>(
        &'life0 self,
        hash: TxHash
    ) -> Pin<Box<dyn Future<Output = TransportResult<Transaction>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_logs<'life0, 'async_trait>(
        &'life0 self,
        filter: Filter
    ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<Log>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_accounts<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<Address>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_gas_price<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = TransportResult<U256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_transaction_receipt<'life0, 'async_trait>(
        &'life0 self,
        hash: TxHash
    ) -> Pin<Box<dyn Future<Output = TransportResult<Option<TransactionReceipt>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_fee_history<'life0, 'life1, 'async_trait>(
        &'life0 self,
        block_count: U256,
        last_block: BlockNumberOrTag,
        reward_percentiles: &'life1 [f64]
    ) -> Pin<Box<dyn Future<Output = TransportResult<FeeHistory>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_block_receipts<'life0, 'async_trait>(
        &'life0 self,
        block: BlockNumberOrTag
    ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Vec<TransactionReceipt>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_uncle<'life0, 'async_trait>(
        &'life0 self,
        tag: BlockId,
        idx: U64
    ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Block>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn syncing<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = TransportResult<SyncStatus>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn call<'life0, 'async_trait>(
        &'life0 self,
        tx: CallRequest,
        block: Option<BlockId>
    ) -> Pin<Box<dyn Future<Output = TransportResult<Bytes>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn call_with_overrides<'life0, 'async_trait>(
        &'life0 self,
        tx: CallRequest,
        block: Option<BlockId>,
        state: StateOverride
    ) -> Pin<Box<dyn Future<Output = TransportResult<Bytes>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn estimate_gas<'life0, 'async_trait>(
        &'life0 self,
        tx: CallRequest,
        block: Option<BlockId>
    ) -> Pin<Box<dyn Future<Output = TransportResult<U256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_raw_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx: Bytes
    ) -> Pin<Box<dyn Future<Output = TransportResult<TxHash>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn estimate_eip1559_fees<'life0, 'async_trait>(
        &'life0 self,
        estimator: Option<EstimatorFunction>
    ) -> Pin<Box<dyn Future<Output = TransportResult<(U256, U256)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_proof<'life0, 'async_trait>(
        &'life0 self,
        address: Address,
        keys: Vec<StorageKey>,
        block: Option<BlockId>
    ) -> Pin<Box<dyn Future<Output = TransportResult<EIP1186AccountProofResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_access_list<'life0, 'async_trait>(
        &'life0 self,
        request: CallRequest,
        block: Option<BlockId>
    ) -> Pin<Box<dyn Future<Output = TransportResult<AccessListWithGasUsed>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_transaction<'life0, 'async_trait>(
        &'life0 self,
        hash: TxHash
    ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<LocalizedTransactionTrace>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn debug_trace_transaction<'life0, 'async_trait>(
        &'life0 self,
        hash: TxHash,
        trace_options: GethDebugTracingOptions
    ) -> Pin<Box<dyn Future<Output = TransportResult<GethTrace>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_block<'life0, 'async_trait>(
        &'life0 self,
        block: BlockNumberOrTag
    ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<LocalizedTransactionTrace>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn raw_request<'life0, 'async_trait, P, R>(
        &'life0 self,
        method: &'static str,
        params: P
    ) -> Pin<Box<dyn Future<Output = TransportResult<R>> + Send + 'async_trait>>
       where P: Serialize + Send + Sync + Clone + 'async_trait,
             R: Serialize + DeserializeOwned + Send + Sync + Unpin + 'static + 'async_trait,
             Self: Sync + 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn get_block<'life0, 'async_trait>(
        &'life0 self,
        id: BlockId,
        full: bool
    ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Block>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }

}

Temporary Provider trait to be used until the new Provider trait with the Network abstraction is stable. Once the new Provider trait is stable, this trait will be removed.

Required Methods

fn get_transaction_count<'life0, 'async_trait>( &'life0 self, address: Address, tag: Option<BlockId> ) -> Pin<Box<dyn Future<Output = TransportResult<U256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the transaction count of the corresponding address.

fn get_block_number<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = TransportResult<u64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the last block number available.

fn get_balance<'life0, 'async_trait>( &'life0 self, address: Address, tag: Option<BlockId> ) -> Pin<Box<dyn Future<Output = TransportResult<U256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the balance of the account at the specified tag, which defaults to latest.

fn get_block_by_hash<'life0, 'async_trait>( &'life0 self, hash: BlockHash, full: bool ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Block>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets a block by its BlockHash, with full transactions or only hashes.

fn get_block_by_number<'life0, 'async_trait>( &'life0 self, number: BlockNumberOrTag, full: bool ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Block>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets a block by BlockNumberOrTag, with full transactions or only hashes.

fn get_client_version<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = TransportResult<String>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the client version of the chain client.

fn get_chain_id<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = TransportResult<U64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the chain ID.

fn get_net_version<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = TransportResult<U64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the network ID. Same as eth_chainId.

fn get_storage_at<'life0, 'async_trait>( &'life0 self, address: Address, key: U256, tag: Option<BlockId> ) -> Pin<Box<dyn Future<Output = TransportResult<StorageValue>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the specified storage value from Address.

fn get_code_at<'life0, 'async_trait>( &'life0 self, address: Address, tag: BlockId ) -> Pin<Box<dyn Future<Output = TransportResult<Bytes>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the bytecode located at the corresponding Address.

fn get_transaction_by_hash<'life0, 'async_trait>( &'life0 self, hash: TxHash ) -> Pin<Box<dyn Future<Output = TransportResult<Transaction>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets a Transaction by its TxHash.

fn get_logs<'life0, 'async_trait>( &'life0 self, filter: Filter ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<Log>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves a Vec<Log> with the given Filter.

fn get_accounts<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<Address>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the accounts in the remote node. This is usually empty unless you’re using a local node.

fn get_gas_price<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = TransportResult<U256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the current gas price.

fn get_transaction_receipt<'life0, 'async_trait>( &'life0 self, hash: TxHash ) -> Pin<Box<dyn Future<Output = TransportResult<Option<TransactionReceipt>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets a TransactionReceipt if it exists, by its TxHash.

fn get_fee_history<'life0, 'life1, 'async_trait>( &'life0 self, block_count: U256, last_block: BlockNumberOrTag, reward_percentiles: &'life1 [f64] ) -> Pin<Box<dyn Future<Output = TransportResult<FeeHistory>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Returns a collection of historical gas information FeeHistory which can be used to calculate the EIP1559 fields maxFeePerGas and maxPriorityFeePerGas.

fn get_block_receipts<'life0, 'async_trait>( &'life0 self, block: BlockNumberOrTag ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Vec<TransactionReceipt>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets the selected block BlockNumberOrTag receipts.

fn get_uncle<'life0, 'async_trait>( &'life0 self, tag: BlockId, idx: U64 ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Block>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets an uncle block through the tag BlockId and index U64.

fn syncing<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = TransportResult<SyncStatus>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets syncing info.

fn call<'life0, 'async_trait>( &'life0 self, tx: CallRequest, block: Option<BlockId> ) -> Pin<Box<dyn Future<Output = TransportResult<Bytes>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Execute a smart contract call with CallRequest without publishing a transaction.

fn call_with_overrides<'life0, 'async_trait>( &'life0 self, tx: CallRequest, block: Option<BlockId>, state: StateOverride ) -> Pin<Box<dyn Future<Output = TransportResult<Bytes>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Execute a smart contract call with CallRequest and state overrides, without publishing a transaction.

Note

Not all client implementations support state overrides.

fn estimate_gas<'life0, 'async_trait>( &'life0 self, tx: CallRequest, block: Option<BlockId> ) -> Pin<Box<dyn Future<Output = TransportResult<U256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Estimate the gas needed for a transaction.

fn send_raw_transaction<'life0, 'async_trait>( &'life0 self, tx: Bytes ) -> Pin<Box<dyn Future<Output = TransportResult<TxHash>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Sends an already-signed transaction.

fn estimate_eip1559_fees<'life0, 'async_trait>( &'life0 self, estimator: Option<EstimatorFunction> ) -> Pin<Box<dyn Future<Output = TransportResult<(U256, U256)>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Estimates the EIP1559 maxFeePerGas and maxPriorityFeePerGas fields. Receives an optional EstimatorFunction that can be used to modify how to estimate these fees.

fn get_proof<'life0, 'async_trait>( &'life0 self, address: Address, keys: Vec<StorageKey>, block: Option<BlockId> ) -> Pin<Box<dyn Future<Output = TransportResult<EIP1186AccountProofResponse>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn create_access_list<'life0, 'async_trait>( &'life0 self, request: CallRequest, block: Option<BlockId> ) -> Pin<Box<dyn Future<Output = TransportResult<AccessListWithGasUsed>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn trace_transaction<'life0, 'async_trait>( &'life0 self, hash: TxHash ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<LocalizedTransactionTrace>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Parity trace transaction.

fn debug_trace_transaction<'life0, 'async_trait>( &'life0 self, hash: TxHash, trace_options: GethDebugTracingOptions ) -> Pin<Box<dyn Future<Output = TransportResult<GethTrace>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn trace_block<'life0, 'async_trait>( &'life0 self, block: BlockNumberOrTag ) -> Pin<Box<dyn Future<Output = TransportResult<Vec<LocalizedTransactionTrace>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn raw_request<'life0, 'async_trait, P, R>( &'life0 self, method: &'static str, params: P ) -> Pin<Box<dyn Future<Output = TransportResult<R>> + Send + 'async_trait>>
where P: Serialize + Send + Sync + Clone + 'async_trait, R: Serialize + DeserializeOwned + Send + Sync + Unpin + 'static + 'async_trait, Self: Sync + 'async_trait, 'life0: 'async_trait,

Provided Methods

fn get_block<'life0, 'async_trait>( &'life0 self, id: BlockId, full: bool ) -> Pin<Box<dyn Future<Output = TransportResult<Option<Block>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gets a block by either its hash, tag, or number, with full transactions or only hashes.

Object Safety

This trait is not object safe.