Struct alloy_providers::ProviderBuilder

pub struct ProviderBuilder<L, N = (), T = ()> { /* private fields */ }

A builder for constructing a Provider from various layers.

This type is similar to tower::ServiceBuilder, with extra complication around maintaining the network and transport types.

Implementations

impl<L, N, T> ProviderBuilder<L, N, T>

pub fn layer<Inner>(self, layer: Inner) -> ProviderBuilder<Stack<T, Inner, L>>

Add a layer to the stack being built. This is similar to tower::ServiceBuilder::layer.

Note:

Layers are added in outer-to-inner order, as in tower::ServiceBuilder. The first layer added will be the first to see the request.

pub fn network<Net: Network>(self) -> ProviderBuilder<L, Net, T>

Change the network.

By default, the network is invalid, and contains the unit type (). This method MUST be called before the provider is built. The client and provider methods only exist when the network is valid.

builder.network::<Arbitrum>()

pub fn client(self, client: RpcClient<T>) -> L::Provider
where L: ProviderLayer<NetworkRpcClient<N, T>, N, T>, T: Transport + Clone, N: Network,

Finish the layer stack by providing a root RpcClient, outputting the final Provider type with all stack components.

This is a convenience function for ProviderBuilder::provider<NetworkRpcClient>.

pub fn provider<P>(self, provider: P) -> L::Provider
where L: ProviderLayer<P, N, T>, P: Provider<N, T>, T: Transport, N: Network,

Finish the layer stack by providing a root Provider, outputting the final Provider type with all stack components.