Module alloy_providers::utils

Provider-related utilities.

Constants

EIP1559_FEE_ESTIMATION_DEFAULT_PRIORITY_FEE

The default max priority fee per gas, used in case the base fee is within a threshold.

EIP1559_FEE_ESTIMATION_PAST_BLOCKS

The number of blocks from the past for which the fee rewards are fetched for fee estimation.

EIP1559_FEE_ESTIMATION_PRIORITY_FEE_TRIGGER

The threshold for base fee below which we use the default priority fee, and beyond which we estimate an appropriate value for priority fee.

EIP1559_FEE_ESTIMATION_REWARD_PERCENTILE

The default percentile of gas premiums that are fetched for fee estimation.

EIP1559_FEE_ESTIMATION_THRESHOLD_MAX_CHANGE

The threshold max change/difference (in %) at which we will ignore the fee history values under it.

Functions

eip1559_default_estimator

The default EIP-1559 fee estimator which is based on the work by MyCrypto

Type Aliases

EstimatorFunction

An estimator function for EIP1559 fees.